Changed 'build-cjs' output dir to 'cjs' folder
To avoid committing files by mistake we have to add them to .gitignore. It's less difficult
when we can ignore the 'umd' and 'cjs' folders, instead of polluting the root dir with compiled files.

In addition, we no longer include all default files, which was making CONTIRBUTING.md part of the tarball downloaded by NPM on each install.